store example
