+++ examples/store/src/text_input.rs
@@ -1,29 +1,28 @@
 use yew::prelude::*;
 
-pub struct TextInput {
-    link: ComponentLink<Self>,
-    text: String,
-    props: TextInputProperties,
-}
-
-pub enum TextInputMsg {
+pub enum Msg {
     SetText(String),
     Submit,
-    None,
 }
 
 #[derive(Properties, Clone, PartialEq)]
-pub struct TextInputProperties {
+pub struct Props {
     pub value: String,
     pub onsubmit: Callback<String>,
 }
 
+pub struct TextInput {
+    link: ComponentLink<Self>,
+    text: String,
+    props: Props,
+}
+
 impl Component for TextInput {
-    type Message = TextInputMsg;
-    type Properties = TextInputProperties;
+    type Message = Msg;
+    type Properties = Props;
 
     fn create(props: Self::Properties, link: ComponentLink<Self>) -> Self {
-        TextInput {
+        Self {
             link,
             text: props.value.clone(),
             props,
@@ -32,14 +31,16 @@ impl Component for TextInput {
 
     fn update(&mut self, msg: Self::Message) -> ShouldRender {
         match msg {
-            TextInputMsg::SetText(text) => self.text = text,
-            TextInputMsg::Submit => {
+            Msg::SetText(text) => {
+                self.text = text;
+                true
+            }
+            Msg::Submit => {
                 let text = std::mem::replace(&mut self.text, self.props.value.clone());
                 self.props.onsubmit.emit(text);
+                true
             }
-            TextInputMsg::None => return false,
         }
-        true
     }
 
     fn change(&mut self, props: Self::Properties) -> ShouldRender {
@@ -57,10 +58,10 @@ impl Component for TextInput {
             <input
                 type="text"
                 value=&self.text
-                oninput=self.link.callback(|e: InputData| TextInputMsg::SetText(e.value))
-                onkeydown=self.link.callback(move |e: KeyboardEvent| {
+                oninput=self.link.callback(|e: InputData| Msg::SetText(e.value))
+                onkeydown=self.link.batch_callback(move |e: KeyboardEvent| {
                     e.stop_propagation();
-                    if e.key() == "Enter" { TextInputMsg::Submit } else { TextInputMsg::None }
+                    if e.key() == "Enter" { Some(Msg::Submit) } else { None }
                 })
             />
         }
