--- examples/textarea/src/main.rs
@@ -1,3 +1,0 @@
-fn main() {
-    yew::start_app::<textarea::Model>();
-}
